requirements.txt
accelerate==1.1.1
captum==0.7.0
colorama==0.4.6
datasets==2.14.5
deepspeed==0.15.4
flash_attn==2.2.1
matplotlib==3.5.3
nltk==3.9.1
numpy==2.1.2
openai==1.52.2
pandas==2.0.3
paramiko==3.5.0
peft==0.6.0
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
rich==13.9.3
scikit_learn==1.3.0
scipy==1.14.1
seaborn==0.13.2
sentence_transformers==3.1.1
spacy==3.7.5
torch==2.4.0
tqdm==4.66.1
transformers==4.45.1
trl==0.7.1
typing_extensions==4.12.2
vllm==0.6.3.post1